ai_resume_analyzer/app/requirements.txt
Authlib==1.3.1
Flask==3.1.2
itsdangerous==2.2.0
pymongo==4.16.0
python-dotenv==1.2.1
pytz==2024.1
requests==2.32.5
Werkzeug==3.1.5
PyMuPDF==1.24.0
python-docx==1.1.0
sentence-transformers==2.7.0
xgboost==2.0.3
